Various fixes/clean-ups to v6fs
-- Backported a fix from original v7 code that corrects a bug
where the inode for the parent directory is left in a locked
state whenever the system runs out of inodes while attempting
to create a new file.

-- Fixed a bug (also present in original v7 code) that resulted in
a leaked inode whenever there was no room to create the directory
entry for a new file.

-- Defend against compilation on big-endian systems (which is
consciously unsupported).

-- Fix up parsing for fssize and fsoffset arguments.

-- Force fuse process into foreground whenever --debug option given.

+++ ancient-src/v6/dmr/bio.c
@@ -303,7 +303,7 @@ void
 notavail(struct buf *bp)
 {
 	register struct buf *rbp;
-	register int sps;
+	register int16_t sps;
 
 	rbp = bp;
 	sps = PS->integ;
